Version 1.0 Commit
Modified code for parallax effect so we don't have to keep updating our CSS whenever we add new content.

All body backgrounds will be centered and start at 0px. wScroll keeps track of the distance between the current window view and the top of the page (0). Scrolling the window triggers our event listener to update the background position of the body element.